backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for the bailee-ann website"
authors = [ "Anis Golriz"]
package-mode= false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "3.1.0"
Flask-JWT-Extended = "4.7.1"
Flask-Login = "0.6.3"
Flask-Migrate = "4.1.0"
Flask-SQLAlchemy = "3.1.1"
alembic = "1.15.1"
click = "8.1.8"
greenlet = "3.1.1"
itsdangerous = "2.2.0"
Mako = "1.3.9"
MarkupSafe = "3.0.2"
psycopg2 = "2.9.10"
PyMySQL = "1.1.1"
python-dotenv = "1.0.1"
SQLAlchemy = "2.0.38"
typing_extensions = "4.12.2"
Werkzeug = "3.1.3"
pytest = "^8.4.1"
requests = "^2.32.4"
flask-cors = "^6.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

   
[tool.isort]
profile = "black"
